refactor(pkg): renamed internal file structure
This will provide better bundling in v3 and just better organization overall. For now, this is not a breaking change. External usage remains the same.
