+++ packages/pilot/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@waveplay/pilot",
-	"version": "0.0.0-canary-20221109054532",
+	"version": "0.0.0-canary-20221109062830",
 	"private": false,
 	"description": "NextJS-like routing for React Native",
 	"repository": {
