+++ packages/pilot/src/client/types.ts
@@ -4,8 +4,8 @@
 import type { GetServerSideProps, GetStaticProps } from 'next';
 import type { NextRouter } from 'next/router';
 import type { ComponentProps, ComponentType } from 'react';
-import { Pilot } from './pilot';
-import { DataMap } from './_internal/types';
+import { Pilot } from './core/pilot';
+import { DataMap } from '../_internal/types';
 
 export interface Logger {
 	debug: (...args: any[]) => void
