--- packages/pilot/src/index.ts
@@ -1,13 +1,0 @@
-/**
- * © 2022 WavePlay <dev@waveplay.com>
- */
-
-// Export all non-internal, uh... exports ÒwÓ
-export * from './area';
-export * from './link';
-export * from './pilot';
-export * from './renderer';
-export * from './radix-router';
-export * from './regex-router';
-export * from './types';
-export * from './use-pilot';
